Change to the default distance parameters: gap_penalty_cdr3_region going from 8 to 12.
We've seen evidence that increasing this gap penalty improves performance slightly.
Also, with the factor of 3 applied to amino-acid mismatches in the CDR3, the new value
of 12 means that we are using the same gap penalty in the V and CDR3 regions, which makes
things simpler and requires fewer overall parameters.

rerun_paper_analysis.py still uses the original parameters from the paper.